requirements.txt
ariadne==0.11.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
graphql-core==3.0.5
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
SQLAlchemy==1.3.22
starlette==0.13.4
typing-extensions==3.7.4.2
Werkzeug==1.0.1